SaveReminderFragment Explain
-Get the view model this time as a single to be shared with the another fragment
- Determines whether the app has the appropriate permissions across Android 10+ and all otherAndroid versions.

The permissionsArray contains the permissions that are going to be requested. Initially, add ACCESS_FINE_LOCATION since that will be needed on all API levels.
 you will need a resultCode. The code will be different depending on if the device is running Q or later and will inform
//     us if you need to check for one permission (fine location) or multiple permissions (fine and background location) when the user returns from the
permission request screen. Add a when statement to check the version running and assign result code to
     REQUEST_FOREGROUND_AND_BACKGROUND_PERMISSION_RESULT_CODE if the device is running Q or later and
REQUEST_FOREGROUND_ONLY_PERMISSIONS_REQUEST_CODE if not.

check if these  permissions are granted or not
ACCESS_FINE_LOCATION
ACCESS_BACKGROUND_LOCATION

In all cases, we need to have location permission.  On Android 10+ (Q) we need to have background permission as well.

-Permissions can be denied in a few ways:
If the grantResults array is empty, then the interaction was interrupted and the permission request was canceled.

If the grantResults array’s value at the LOCATION_PERMISSION_INDEX has a PERMISSION_DENIED it means that the user denied foreground permissions.

If the request code equals REQUEST_FOREGROUND_AND_BACKGROUND_PERMISSION_RESULT_CODE and the BACKGROUND_LOCATION_PERMISSION_INDEX is denied it means that the device is running API 29 or above and that background permissions were denied.

 Replace the code below in the onActivityResult() method. After the user

chooses whether to accept or deny device location permissions, this checks